requirements.txt
captum==0.7.0
grad_cam==1.5.2
hdbscan==0.8.37
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.10.0.82
pandas==2.0.3
Pillow==9.4.0
pykalman==0.9.7
scikit_learn==1.3.2
screeninfo==0.8.1
torch==2.2.0
torchvision==0.17.0
tensorboard